ci: add license attributions to Dockerfile
While we already have THIRD-PARTY-LICENSES.md, Amazon legal required
that the output be from go-licenses, which lists all licenses from all
of our dependencies.

This change adds license attributions to the Dockerfile in CI and
uploads them in the release image.

Also with this change, a tag push is enough to create a draft release
with the needed artifacts. From there they can be downloaded maunally.
We also only build Linux variants since that is all we test with.

Signed-off-by: David Son <davbson@amazon.com>